+++ packages/plugin-packages/mashroom-portal-remote-app-registry-k8s/src/js/jobs/mashroom-bootstrap-background-job.ts
@@ -7,15 +7,15 @@ import type {MashroomBackgroundJobPluginBootstrapFunction} from '@mashroom/mashr
 
 const bootstrap: MashroomBackgroundJobPluginBootstrapFunction = (pluginName, pluginConfig, pluginContextHolder) => {
     const pluginContext = pluginContextHolder.getPluginContext();
-    const {k8sNamespaces, socketTimeoutSec, refreshIntervalSec, serviceNameFilter, accessViaClusterIP} = pluginConfig;
+    const {serviceNameFilter, k8sNamespacesLabelSelector, k8sNamespaces, socketTimeoutSec, refreshIntervalSec, accessViaClusterIP} = pluginConfig;
 
     context.serviceNameFilter = serviceNameFilter;
 
     const kubernetesConnector = process.env.DUMMY_K8S_CONNECTOR ? new DummyKubernetesConnector() : new KubernetesConnector();
 
     const backgroundJob = new ScanK8SPortalRemoteAppsBackgroundJob(
-        k8sNamespaces, serviceNameFilter, socketTimeoutSec,
-        refreshIntervalSec, accessViaClusterIP,
+        serviceNameFilter, k8sNamespacesLabelSelector, k8sNamespaces,
+        socketTimeoutSec, refreshIntervalSec, accessViaClusterIP,
         pluginContext.serverConfig.externalPluginConfigFileNames,
         kubernetesConnector, pluginContext.loggerFactory);