+++ packages/plugin-packages/mashroom-portal-remote-app-registry-k8s/src/js/k8s/KubernetesConnector.ts
@@ -1,5 +1,5 @@
 
-import {CoreV1Api, KubeConfig, V1ServiceList} from '@kubernetes/client-node';
+import {CoreV1Api, KubeConfig, V1NamespaceList, V1ServiceList} from '@kubernetes/client-node';
 
 import type {KubernetesConnector as KubernetesConnectorType} from '../../../type-definitions';
 
@@ -7,7 +7,21 @@ export default class KubernetesConnector implements KubernetesConnectorType {
 
     private _k8sApi?: CoreV1Api;
 
-    async listNamespaceServices(namespace: string): Promise<V1ServiceList> {
+    constructor(private test?: boolean) {
+    }
+
+    async getNamespacesByLabel(labelSelector: string): Promise<V1NamespaceList> {
+        if (!this._k8sApi) {
+            this.init();
+        }
+        if (!this._k8sApi) {
+            throw new Error('No k8s client found.');
+        }
+        const result = await this._k8sApi.listNamespace(undefined, undefined, undefined, undefined, labelSelector);
+        return result.body;
+    }
+
+    async getNamespaceServices(namespace: string): Promise<V1ServiceList> {
         if (!this._k8sApi) {
             this.init();
         }
@@ -20,8 +34,13 @@ export default class KubernetesConnector implements KubernetesConnectorType {
 
     private init(): void {
         const k8sClient = new KubeConfig();
-        // This only works if the Portal runs within a Kubernetes Pod with a valid service account attached
-        k8sClient.loadFromCluster();
+        if (!this.test) {
+            // This only works if the Portal runs within a Kubernetes Pod with a valid service account attached
+            k8sClient.loadFromCluster();
+        } else {
+            // This uses the local kubectl config (only for test purposes!)
+            k8sClient.loadFromDefault();
+        }
         this._k8sApi = k8sClient.makeApiClient(CoreV1Api);
     }